refactor: 피드백 적용
