+++ src/main/resources/application.properties
@@ -3,6 +3,6 @@ spring.h2.console.enabled=false
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:13306/wootecochess?useSSL=false&serverTimezone=UTC&useUnicode=true&characterEncoding=utf8
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=root
 
 logging.level.org.springframework.jdbc.core.JdbcTemplate=debug
