--- src/main/resources/templates/css/board.css
@@ -1,140 +1,0 @@
-body {
-    background-position: center;
-    background-size: cover;
-}
-
-#app {
-    display: flex;
-    margin: 0;
-    padding: 0;
-    border: 0;
-    height: 773px;
-}
-
-.turn-number {
-    display: none;
-    width: 1;
-}
-
-.board {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin: auto 20px;
-    margin-left: 0;
-    height: 100%;
-    width: 63%;
-}
-
-.row {
-    display: flex;
-    text-align: center;
-    margin: 0 auto;
-    width: 85%;
-}
-
-.cell {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-}
-
-.white-cell {
-    background-color: white;
-}
-
-.black-cell {
-    background-color: black;
-}
-
-.selected-cell {
-    background-color: #FFFACD;
-    opacity: 0.9;
-    border: 0.5px solid gray;
-}
-
-.cell > img {
-    width: 100%;
-    margin: 0 auto;
-}
-
-.y {
-    background-color: #FFFACD;
-    opacity: 0.9;
-    border: 0.5px solid gray;
-}
-
-.side {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    height: 100%;
-    width: 37%;
-}
-
-.button {
-    height: 90px;
-    width: 100%;
-    margin: 2px 0;
-}
-
-.user {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    background-color: seashell;
-    border-radius: 10px;
-}
-
-.user > img {
-    flex: 1;
-    height: 150px;
-    margin: 0 auto;
-}
-
-.game {
-    display: flex;
-}
-
-.user-status {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    width: 100%;
-    height: 150px;
-    margin: 5px auto;
-}
-
-.user-status .user-name {
-    margin: 0;
-    font-size: 50px;
-    font-family: fantasy;
-    text-align: center;
-    flex: 1;
-}
-
-.user-status .turn {
-    margin: 0;
-    font-size: 50px;
-    font-family: fantasy;
-    color: hotpink;
-    flex: 1;
-}
-
-.user-status .score {
-    text-align: center;
-    font-family: fantasy;
-    font-size: 30px;
-}
-
-input[type="submit"] {
-    width: 100%;
-    height: 100%;
-    border-radius: 10px;
-    background-color: rosybrown;
-    font-family: Arial, Helvetica, Sans-serif;
-    font-size: 50px;
-}
