--- src/main/resources/templates/css/main.css
@@ -1,85 +1,0 @@
-body {
-    background-position: center;
-    background-size: cover;
-}
-
-#app {
-    display: flex;
-    margin: 0;
-    padding: 0;
-    border: 0;
-    height: 773px;
-}
-
-.board {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    height: 100%;
-    width: 63%;
-}
-
-.row {
-    display: flex;
-    text-align: center;
-    margin: 0 auto;
-    width: 85%;
-}
-
-.cell {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-}
-
-.white-cell {
-    background-color: white;
-}
-
-.black-cell {
-    background-color: black;
-}
-
-.cell > img {
-    width: 100%;
-    margin: 0 auto;
-}
-
-.start {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    width: 37%;
-    height: 100%;
-}
-
-form {
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 50%;
-}
-
-.users {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.user {
-    margin: 10px auto;
-    font-size: 30px;
-    font-family: Georgia, 'Times New Roman', Times, serif;
-    color: black;
-}
-
-input[type="submit"] {
-    width: 100%;
-    height: 100px;
-    border-radius: 10px;
-    background-color: rosybrown;
-    font-family: Arial, Helvetica, Sans-serif;
-    font-size: 50px;
-}
