+++ src/test/java/wooteco/chess/controller/ChessControllerTest.java
@@ -1,0 +1,7 @@
+package wooteco.chess.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ChessControllerTest {
+
+}
